Add new level and add gravity feature

+++ src/com/soft/entity/Entity.java
@@ -6,7 +6,7 @@ public abstract class Entity {
 
     protected int x,y;
     protected int width, height;
-    protected int velocityX, velocityY;
+    protected float velocityX, velocityY;
 
     public Entity(int x, int y, int width, int height) {
         this.x = x;
